requirements.txt
# Web Framework
fastapi>=0.115.0
uvicorn>=0.32.0

# OpenAI
openai>=1.50.0

# Data Processing
numpy>=1.26.0
pydantic-settings>=2.0.0

# Environment
python-dotenv>=1.0.0

# HTTP Client
httpx[socks]>=0.27.0

# LangGraph / LangChain
langgraph>=0.2.0
langchain-core>=0.3.0

# Vector Store
chromadb>=0.4.0

# Resilience
tenacity>=8.2.0
cachetools>=5.3.0

# Database
aiosqlite>=0.19.0
